+++ scripts/utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+Hilfsfunktionen für die Dota Analyse Pipeline
+"""